# include overflows in symbology