--- a/datamodel/app/symbology_functions.sql
+++ b/datamodel/app/symbology_functions.sql
@@ -23,14 +23,24 @@ SET
   _usage_current = usage_current,
   _status = status
 FROM(
-  SELECT DISTINCT ON (ne.obj_id) ne.obj_id AS ne_obj_id,
-      COALESCE(first_value(CH_from.function_hierarchic) OVER w, first_value(CH_to.function_hierarchic) OVER w) AS function_hierarchic,
-      COALESCE(first_value(CH_from.usage_current) OVER w, first_value(CH_to.usage_current) OVER w) AS usage_current,
-      COALESCE(first_value(ws_from.status) OVER w, first_value(ws_to.status) OVER w) AS status,
+  SELECT DISTINCT ON (wn.obj_id) wn.obj_id AS wn_obj_id,
+      COALESCE(first_value(CH_from.function_hierarchic) OVER w
+              , first_value(CH_to.function_hierarchic) OVER w
+              , first_value(CH_ov_from.function_hierarchic) OVER w
+              , first_value(CH_ov_to.function_hierarchic) OVER w) AS function_hierarchic,
+      COALESCE(first_value(CH_from.usage_current) OVER w
+              , first_value(CH_to.usage_current) OVER w
+              , first_value(CH_ov_from.usage_current) OVER w
+              , first_value(CH_ov_to.usage_current) OVER w) AS usage_current,
+      COALESCE(first_value(ws_from.status) OVER w
+             , first_value(ws_to.status) OVER w
+             , first_value(ws_ov_from.status) OVER w
+             , first_value(ws_ov_to.status) OVER w) AS status,
       rank() OVER w AS hierarchy_rank
     FROM
-      tww_od.wastewater_networkelement ne
-      LEFT JOIN tww_od.reach_point rp ON ne.obj_id = rp.fk_wastewater_networkelement
+      tww_od.wastewater_node wn
+	  
+	  LEFT JOIN tww_od.reach_point                 rp                   ON wn.obj_id = rp.fk_wastewater_networkelement
       LEFT JOIN tww_od.reach                       re_from           	ON re_from.fk_reach_point_from = rp.obj_id
       LEFT JOIN tww_od.wastewater_networkelement   ne_from           	ON ne_from.obj_id = re_from.obj_id
       LEFT JOIN tww_od.channel                     CH_from           	ON CH_from.obj_id = ne_from.fk_wastewater_structure
@@ -44,11 +54,40 @@ FROM(
       LEFT JOIN tww_od.wastewater_structure        ws_to           	ON ws_to.obj_id = ne_to.fk_wastewater_structure
       LEFT JOIN tww_vl.channel_function_hierarchic vl_fct_hier_to 	ON CH_to.function_hierarchic = vl_fct_hier_to.code
       LEFT JOIN tww_vl.channel_usage_current       vl_usg_curr_to 	ON CH_to.usage_current = vl_usg_curr_to.code
-    WHERE _all OR ne.obj_id = _obj_id
-      WINDOW w AS ( PARTITION BY ne.obj_id ORDER BY vl_fct_hier_from.order_fct_hierarchic ASC NULLS LAST, vl_fct_hier_to.order_fct_hierarchic ASC NULLS LAST,
-                                vl_usg_curr_from.order_usage_current ASC NULLS LAST, vl_usg_curr_to.order_usage_current ASC NULLS LAST ROWS BETWEEN UNBOUNDED PRECEDING AND UNBOUNDED FOLLOWING)
+	  
+	  -- Update by Overflow
+	  LEFT JOIN tww_od.overflow                    ov         ON ov.fk_overflow_to=wn.obj_id
+      LEFT JOIN tww_od.wastewater_networkelement   ne_ov      ON ne_ov.obj_id = ov.fk_wastewater_node
+
+      LEFT JOIN tww_od.reach_point                 rp_ov_from          ON ne_ov.obj_id = rp_ov_from.fk_wastewater_networkelement
+      LEFT JOIN tww_od.reach                       re_ov_from          ON re_ov_from.fk_reach_point_from = rp_ov_from.obj_id
+      LEFT JOIN tww_od.wastewater_networkelement   ne_ov_from          ON ne_ov_from.obj_id = re_ov_from.obj_id
+      LEFT JOIN tww_od.channel                     CH_ov_from          ON CH_ov_from.obj_id = ne_ov_from.fk_wastewater_structure
+      LEFT JOIN tww_od.wastewater_structure        ws_ov_from          ON ws_ov_from.obj_id = ne_ov_from.fk_wastewater_structure
+      LEFT JOIN tww_vl.channel_function_hierarchic vl_fct_hier_ov_from	ON CH_ov_from.function_hierarchic = vl_fct_hier_ov_from.code
+      LEFT JOIN tww_vl.channel_usage_current       vl_usg_curr_ov_from	ON CH_ov_from.usage_current = vl_usg_curr_ov_from.code
+
+      LEFT JOIN tww_od.reach_point                 rp_ov_to            ON ne_ov.obj_id = rp_ov_to.fk_wastewater_networkelement
+      LEFT JOIN tww_od.reach                       re_ov_to           	ON re_ov_to.fk_reach_point_to = rp_ov_to.obj_id
+      LEFT JOIN tww_od.wastewater_networkelement   ne_ov_to           	ON ne_ov_to.obj_id = re_ov_to.obj_id
+      LEFT JOIN tww_od.channel                     CH_ov_to           	ON CH_ov_to.obj_id = ne_ov_to.fk_wastewater_structure
+      LEFT JOIN tww_od.wastewater_structure        ws_ov_to           	ON ws_ov_to.obj_id = ne_ov_to.fk_wastewater_structure
+      LEFT JOIN tww_vl.channel_function_hierarchic vl_fct_hier_ov_to	  ON CH_ov_to.function_hierarchic = vl_fct_hier_ov_to.code
+      LEFT JOIN tww_vl.channel_usage_current       vl_usg_curr_ov_to	  ON CH_ov_to.usage_current = vl_usg_curr_ov_to.code
+    WHERE _all OR wn.obj_id = _obj_id
+      WINDOW w AS ( PARTITION BY wn.obj_id 
+                    ORDER BY vl_fct_hier_from.order_fct_hierarchic ASC NULLS LAST
+                           , vl_fct_hier_to.order_fct_hierarchic ASC NULLS LAST
+                           , vl_fct_hier_ov_from.order_fct_hierarchic ASC NULLS LAST
+                           , vl_fct_hier_ov_to.order_fct_hierarchic ASC NULLS LAST
+
+                           , vl_usg_curr_from.order_usage_current ASC NULLS LAST
+                           , vl_usg_curr_to.order_usage_current ASC NULLS LAST
+                           , vl_usg_curr_ov_from.order_usage_current ASC NULLS LAST
+                           , vl_usg_curr_ov_to.order_usage_current ASC NULLS LAST
+                    ROWS BETWEEN UNBOUNDED PRECEDING AND UNBOUNDED FOLLOWING)
 ) symbology_ne
-WHERE symbology_ne.ne_obj_id = n.obj_id;
+WHERE symbology_ne.wn_obj_id = n.obj_id;
 
 -- See above
 IF _all THEN