--- a/datamodel/app/symbology_functions.sql
+++ b/datamodel/app/symbology_functions.sql
@@ -32,14 +32,17 @@ FROM(
               , first_value(CH_to.usage_current) OVER w
               , first_value(CH_ov_from.usage_current) OVER w
               , first_value(CH_ov_to.usage_current) OVER w) AS usage_current,
-      COALESCE(first_value(ws_from.status) OVER w
+      COALESCE(first_value(ws_node.status) OVER w
+             , first_value(ws_from.status) OVER w
              , first_value(ws_to.status) OVER w
              , first_value(ws_ov_from.status) OVER w
              , first_value(ws_ov_to.status) OVER w) AS status,
       rank() OVER w AS hierarchy_rank
     FROM
       tww_od.wastewater_node wn
-
+      LEFT JOIN tww_od.wastewater_networkelement   ne                   ON ne.obj_id = wn.obj_id
+	  LEFT JOIN tww_od.wastewater_structure        ws_node              ON ws_node.obj_id = ne.fk_wastewater_structure
+	  
 	  LEFT JOIN tww_od.reach_point                 rp                   ON wn.obj_id = rp.fk_wastewater_networkelement
       LEFT JOIN tww_od.reach                       re_from           	ON re_from.fk_reach_point_from = rp.obj_id
       LEFT JOIN tww_od.wastewater_networkelement   ne_from           	ON ne_from.obj_id = re_from.obj_id