There are several situations (i.E. a stormwater composite tank with a preceding separating structure) where a wastewater structure only has overflows as topological connections. In these situations, we should also set a _usage_current dependent on the incoming overflow. This PR includes them in the updating process for wastewater nodes.

See also https://github.com/QGEP/datamodel/pull/225/files