@ponceta 2024-04-23
It looks good, does it works good too?

Have you tried it on with a data set? My main concern is the performance, if it's green let's go!