@cymed 2024-04-25
> It looks good, does it works good too?
> 
> Have you tried it on with a data set? My main concern is the performance, if it's green let's go!

It works with real data. 
For performance, I have tested the `SELECT DISTINCT ON ...` bit with _all=true on a small db of 2'570 wastewater nodes. 

Before: 
Planning Time: 3.110   ms
Execution Time: 77.668 ms

New: 
Planning Time:   12.957 ms
Execution Time: 129.622 ms

For a single wastewater node (i.e. when altering a wastewater structure)
Before: 
Planning Time: 3.188   ms
Execution Time: 0.216 ms

New: 
Planning Time:   8.382 ms
Execution Time: 0.525 ms

If the performance drop bothers you I suggest to store` _usage_current`, `_status `and `_function_hierarchic ` on tww.overflow and grab it from there