@ponceta 2024-04-29
I'm OK to go with that but I would enjoy a good schema in the doc on how it works and what special structures can be well modelized with that, otherwise it is quite an additionnal complexity in terms of symbology.

I think it is mainly on the overflows "new" mechanics and it would be good if one can achieve it without being a wastewater datamanager certified engineer. (otherwise I'm quite excited to see how it looks in real life examples.)