@cymed 2024-04-30
> I'm OK to go with that but I would enjoy a good schema in the doc on how it works and what special structures can be well modelized with that, otherwise it is quite an additionnal complexity in terms of symbology.
> 
> I think it is mainly on the overflows "new" mechanics and it would be good if one can achieve it without being a wastewater datamanager certified engineer. (otherwise I'm quite excited to see how it looks in real life examples.)

I would not add any additional docs, as the correct topology is defined in the vsa wiki. It is used for cases such as this
![grafik](https://github.com/teksi/wastewater/assets/102035587/de8b8af1-a7dc-461c-93b8-873ebc8266df)
We have an overflow flowing into a cso tank. the tank has an overflow into the water body. Here, the cso tank's wastewater_node did not receive a _usage_current / _status /_function_hierarchic, as there was no topological connection to a reach.  With the PR, the result is

The last pushes improve general performance as there are fewer bulky left joins.