@cymed 2024-04-30
The last pushes improve performance and readability. The only change to the original function is that I included the wn's corresponding ws for status.

All the overflow specific stuff is separated from the other PR, making it much faster as it left joins the nodes onto an overflow instead of left joining the overflow on the node.

> I'm OK to go with that but I would enjoy a good schema in the doc on how it works and what special structures can be well modelized with that, otherwise it is quite an additionnal complexity in terms of symbology.
> 
> I think it is mainly on the overflows "new" mechanics and it would be good if one can achieve it without being a wastewater datamanager certified engineer. (otherwise I'm quite excited to see how it looks in real life examples.)

I don't think we need another docs. It only affects symbology on wastewater nodes for those who follow the vsa wiki. The rest won't notice.

Here is a real-life example, where there is an separating structure is connected to a combined sewer overflow with retention volume , which in turn has an overflow into the water body. Without the PF, the CSO has no usage_current
![grafik](https://github.com/teksi/wastewater/assets/102035587/b5c16ee6-1bb8-4de2-888f-71779cd98f95)