@sjib 2024-04-30
@cymed @ponceta  May be we could include this screenshot in this chapter: https://qgep.github.io/docs/user-guide/How-To/index.html#hydraulic-modeling-of-an-overflow-prank-weir-leapingweir-pump